pyproject.toml
[project]
name = "specweaver"
version = "0.1.0"
description = "Specification-driven development lifecycle tool"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [
    { name = "sbula" },
]

dependencies = [
    "typer[all]>=0.21",
    "google-genai>=1.0.0",
    "pydantic>=2.12",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "ruamel.yaml>=0.18",
    "tomlkit>=0.12.0",
    "jinja2>=3.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-java>=0.23.0",
    "tree-sitter-kotlin>=1.1.0",
    "junitparser>=3.1.2",
    "sarif-tools>=1.0.0",
]

[project.scripts]
sw = "specweaver.interfaces.cli.main:app"

[project.optional-dependencies]
openai = ["openai>=1.5"]
anthropic = ["anthropic>=0.25"]
mistral = ["mistralai>=1.0"]
all-llm = ["openai>=1.5", "anthropic>=0.25", "mistralai>=1.0"]
serve = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.9",
    "markdown>=3.5",
    "bleach>=6.1",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.11",
    "mypy>=1.15",
    "httpx>=0.28",
    "tach",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/specweaver"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-v --tb=short -m 'not live'"
markers = [
    "live: tests that require real external API keys and network access",
]

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
    "C90", # mccabe complexity
]
ignore = [
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["specweaver"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
explicit_package_bases = true
mypy_path = "src"
